fix: update voting power and vesting data together
